+++ wolf-page-builder.php
@@ -89,20 +89,6 @@ final class Wolf_Page_Builder {
 	}
 
 	/**
-	 * Cloning is forbidden.
-	 */
-	public function __clone() {
-		_doing_it_wrong( __FUNCTION__, esc_html__( 'Cheatin&#8217; huh?', '%TEXTDOMAIN%' ), '1.0' );
-	}
-
-	/**
-	 * Unserializing instances of this class is forbidden.
-	 */
-	public function __wakeup() {
-		_doing_it_wrong( __FUNCTION__, esc_html__( 'Cheatin&#8217; huh?', '%TEXTDOMAIN%' ), '1.0' );
-	}
-
-	/**
 	 * %NAME% Constructor.
 	 */
 	public function __construct() {